requirements.txt
pandas
numpy
scipy
scikit-learn
# For Streamlit Cloud, especially if you encounter sqlite version issues:
pysqlite3-binary
streamlit